+++ test/unit/specs/util/EntityToFormMapper.spec.js
@@ -885,15 +885,37 @@ describe('Entity to state mapper', () => {
   describe('Auto generated ( server side) field value validation', () => {
     const data = {}
     const form = EntityToFormMapper.generateForm(schemas.autoIdSchema, data)
-    const field = form.formFields[0]
 
-    it('should map required and visible field attrs to function always evaluating to true', () => {
-      expect(field.type).to.equal('text')
-      expect(field.id).to.equal('id')
+    it('should not map auto, non-visible attribute to field', () => {
+      expect(form.formFields.length).to.equal(0)
+    })
+  })
+
+  describe('MapperMode option', () => {
+    const data = {}
+
+    it('Setting the mapper mode to CREATE should set readonly fields to enabled', () => {
+      const form = EntityToFormMapper.generateForm(schemas.createRowSchema, data, {mapperMode: 'CREATE'})
+      const field = form.formFields[0]
+      expect(field.disabled).to.equal(false)
+      expect(field.readOnly).to.equal(false)
+      expect(field.visible()).to.equal(true)
+    })
+
+    it('Setting the mapper mode to UPDATE should set readonly fields to enabled', () => {
+      const form = EntityToFormMapper.generateForm(schemas.createRowSchema, data, {mapperMode: 'UPDATE'})
+      const field = form.formFields[0]
       expect(field.disabled).to.equal(true)
       expect(field.readOnly).to.equal(true)
-      expect(field.visible()).to.equal(false)
-      expect(field.validate()).to.equal(true)
+      expect(field.visible()).to.equal(true)
+    })
+
+    it('Not setting the mapper mode result is the default UPDATE mode being used', () => {
+      const form = EntityToFormMapper.generateForm(schemas.createRowSchema, data)
+      const field = form.formFields[0]
+      expect(field.disabled).to.equal(true)
+      expect(field.readOnly).to.equal(true)
+      expect(field.visible()).to.equal(true)
     })
   })
 })